ТЗ 14 V1

+++ src/main/java/ru/practicum/shareit/user/service/UserService.java
@@ -6,13 +6,13 @@ import ru.practicum.shareit.user.model.User;
 import java.util.List;
 
 public interface UserService {
-    UserDto getUserById(Integer userId);
+    UserDto getUserById(Long userId);
 
     List<UserDto> findAllUsers();
 
     UserDto createUser(User user);
 
-    UserDto updateUser(User user, Integer userId);
+    UserDto updateUser(User user, Long userId);
 
-    void deleteUser(Integer userId);
+    void deleteUser(Long userId);
 }
